# perf(turbo-rcstr): Fix the fast codepath for RcStr::into_owned()