--- a/turbopack/crates/turbo-rcstr/src/lib.rs
+++ b/turbopack/crates/turbo-rcstr/src/lib.rs
@@ -3,7 +3,7 @@ use std::{
     ffi::OsStr,
     fmt::{Debug, Display},
     hash::{Hash, Hasher},
-    mem::forget,
+    mem::{forget, ManuallyDrop},
     num::NonZeroU8,
     ops::Deref,
     path::{Path, PathBuf},
@@ -96,15 +96,11 @@ impl RcStr {
     pub fn into_owned(self) -> String {
         match self.tag() {
             DYNAMIC_TAG => {
-                let arc = unsafe { dynamic::restore_arc(self.unsafe_data) };
-
-                match Arc::try_unwrap(arc.clone()) {
+                // convert `self` into `arc`
+                let arc = unsafe { dynamic::restore_arc(ManuallyDrop::new(self).unsafe_data) };
+                match Arc::try_unwrap(arc) {
                     Ok(v) => v,
-                    Err(arc) => {
-                        let s = arc.to_string();
-                        forget(arc);
-                        s
-                    }
+                    Err(arc) => arc.to_string(),
                 }
             }
             INLINE_TAG => self.as_str().to_string(),
@@ -298,5 +294,35 @@ impl Drop for RcStr {
         if self.tag() == DYNAMIC_TAG {
             unsafe { drop(dynamic::restore_arc(self.unsafe_data)) }
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::mem::ManuallyDrop;
+
+    use super::*;
+
+    #[test]
+    fn test_refcount() {
+        fn refcount(str: &RcStr) -> usize {
+            assert!(str.tag() == DYNAMIC_TAG);
+            let arc = ManuallyDrop::new(unsafe { dynamic::restore_arc(str.unsafe_data) });
+            triomphe::Arc::count(&arc)
+        }
+
+        let str = RcStr::from("this is a long string that won't be inlined");
+
+        assert_eq!(refcount(&str), 1);
+        assert_eq!(refcount(&str), 1); // refcount should not modify the refcount itself
+
+        let cloned_str = str.clone();
+        assert_eq!(refcount(&str), 2);
+
+        drop(cloned_str);
+        assert_eq!(refcount(&str), 1);
+
+        let _ = str.clone().into_owned();
+        assert_eq!(refcount(&str), 1);
     }
 }