If we uniquely own the only reference to an `RcStr`, `into_owned` should just unwrap the `Arc` and return the inner value instead of cloning it.

It seems this optimization was accidentally broken at some point. `Arc::try_unwrap(arc.clone())` is always going to return the non-unique `Err` value because the `arc.clone()` will always set the refcount >= 2.

This makes `RcStr::into_owned` significantly faster for large strings that we own a unique `Arc` for.

Includes a unit test and a benchmark to prevent regressing.


```
RcStr::map/unique       time:   [17.311 ns 17.539 ns 17.797 ns]
                        change: [-97.932% -97.855% -97.778%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) high mild
  6 (6.00%) high severe

RcStr::map/cloned       time:   [1.0424 µs 1.0529 µs 1.0655 µs]
                        change: [-1.2209% +1.1551% +3.8229%] (p = 0.39 > 0.05)
                        No change in performance detected.
Found 22 outliers among 100 measurements (22.00%)
  9 (9.00%) low severe
  7 (7.00%) low mild
  2 (2.00%) high mild
  4 (4.00%) high severe
```