@ijjk 2025-02-03
# Stats from current PR

<details open>
<summary><strong>Default Build</strong> (Increase detected ⚠️)</summary>

<br/>

<details>
<summary><strong>General</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| buildDuration | 17.8s | 18.2s | ⚠️ +368ms |
| buildDurationCached | 15.1s | 12.7s | N/A |
| nodeModulesSize | 391 MB | 391 MB | N/A |
| nextStartRea..uration (ms) | 423ms | 443ms | N/A |

</details>

<details>
<summary><strong>Client Bundles (main, webpack)</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| 5306-HASH.js gzip | 54 kB | 53.9 kB | N/A |
| 8276.HASH.js gzip | 169 B | 168 B | N/A |
| 8377-HASH.js gzip | 5.46 kB | 5.46 kB | N/A |
| bccd1874-HASH.js gzip | 52.9 kB | 52.9 kB | N/A |
| framework-HASH.js gzip | 57.5 kB | 57.5 kB | N/A |
| main-app-HASH.js gzip | 241 B | 242 B | N/A |
| main-HASH.js gzip | 34.5 kB | 34.4 kB | N/A |
| webpack-HASH.js gzip | 1.71 kB | 1.71 kB | N/A |
| Overall change | 0 B | 0 B | ✓ |

</details>

<details>
<summary><strong>Legacy Client Bundles (polyfills)</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| polyfills-HASH.js gzip | 39.4 kB | 39.4 kB | ✓ |
| Overall change | 39.4 kB | 39.4 kB | ✓ |

</details>

<details>
<summary><strong>Client Pages</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| _app-HASH.js gzip | 193 B | 193 B | ✓ |
| _error-HASH.js gzip | 193 B | 193 B | ✓ |
| amp-HASH.js gzip | 512 B | 510 B | N/A |
| css-HASH.js gzip | 343 B | 342 B | N/A |
| dynamic-HASH.js gzip | 1.84 kB | 1.84 kB | ✓ |
| edge-ssr-HASH.js gzip | 265 B | 265 B | ✓ |
| head-HASH.js gzip | 363 B | 362 B | N/A |
| hooks-HASH.js gzip | 393 B | 392 B | N/A |
| image-HASH.js gzip | 4.59 kB | 4.58 kB | N/A |
| index-HASH.js gzip | 268 B | 268 B | ✓ |
| link-HASH.js gzip | 2.35 kB | 2.35 kB | N/A |
| routerDirect..HASH.js gzip | 328 B | 328 B | ✓ |
| script-HASH.js gzip | 397 B | 397 B | ✓ |
| withRouter-HASH.js gzip | 323 B | 326 B | N/A |
| 1afbb74e6ecf..834.css gzip | 106 B | 106 B | ✓ |
| Overall change | 3.59 kB | 3.59 kB | ✓ |

</details>

<details>
<summary><strong>Client Build Manifests</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| _buildManifest.js gzip | 748 B | 747 B | N/A |
| Overall change | 0 B | 0 B | ✓ |

</details>

<details>
<summary><strong>Rendered Page Sizes</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| index.html gzip | 523 B | 522 B | N/A |
| link.html gzip | 538 B | 537 B | N/A |
| withRouter.html gzip | 519 B | 519 B | ✓ |
| Overall change | 519 B | 519 B | ✓ |

</details>

<details>
<summary><strong>Edge SSR bundle Size</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| edge-ssr.js gzip | 129 kB | 129 kB | N/A |
| page.js gzip | 210 kB | 210 kB | N/A |
| Overall change | 0 B | 0 B | ✓ |

</details>

<details>
<summary><strong>Middleware size</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| middleware-b..fest.js gzip | 670 B | 665 B | N/A |
| middleware-r..fest.js gzip | 155 B | 156 B | N/A |
| middleware.js gzip | 31.3 kB | 31.3 kB | N/A |
| edge-runtime..pack.js gzip | 844 B | 844 B | ✓ |
| Overall change | 844 B | 844 B | ✓ |

</details>

<details>
<summary><strong>Next Runtimes</strong></summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| app-page-exp...dev.js gzip | 386 kB | 386 kB | N/A |
| app-page-exp..prod.js gzip | 132 kB | 132 kB | ✓ |
| app-page-tur..prod.js gzip | 145 kB | 145 kB | ✓ |
| app-page-tur..prod.js gzip | 141 kB | 141 kB | ✓ |
| app-page.run...dev.js gzip | 373 kB | 373 kB | N/A |
| app-page.run..prod.js gzip | 128 kB | 128 kB | ✓ |
| app-route-ex...dev.js gzip | 39.2 kB | 39.2 kB | ✓ |
| app-route-ex..prod.js gzip | 25.6 kB | 25.6 kB | ✓ |
| app-route-tu..prod.js gzip | 25.6 kB | 25.6 kB | ✓ |
| app-route-tu..prod.js gzip | 25.4 kB | 25.4 kB | ✓ |
| app-route.ru...dev.js gzip | 40.8 kB | 40.8 kB | ✓ |
| app-route.ru..prod.js gzip | 25.4 kB | 25.4 kB | ✓ |
| dist_client_...dev.js gzip | 356 B | 356 B | ✓ |
| dist_client_...dev.js gzip | 349 B | 349 B | ✓ |
| pages-api-tu..prod.js gzip | 9.69 kB | 9.69 kB | ✓ |
| pages-api.ru...dev.js gzip | 11.8 kB | 11.8 kB | ✓ |
| pages-api.ru..prod.js gzip | 9.68 kB | 9.68 kB | ✓ |
| pages-turbo...prod.js gzip | 21.9 kB | 21.9 kB | ✓ |
| pages.runtim...dev.js gzip | 31.5 kB | 31.5 kB | ✓ |
| pages.runtim..prod.js gzip | 21.9 kB | 21.9 kB | ✓ |
| server.runti..prod.js gzip | 60.2 kB | 60.2 kB | ✓ |
| Overall change | 895 kB | 895 kB | ✓ |

</details>

<details>
<summary><strong>build cache</strong> Overall increase ⚠️</summary>

|  | vercel/next.js canary  | vercel/next.js bgw/avoid-clone-in-rcstr-into-owned | Change |
| - | - | - | - |
| 0.pack gzip | 2.1 MB | 2.1 MB | ⚠️ +1.89 kB |
| index.pack gzip | 75.3 kB | 75.9 kB | ⚠️ +611 B |
| Overall change | 2.18 MB | 2.18 MB | ⚠️ +2.5 kB |

</details>

<details>
<summary><strong>Diff details</strong></summary>

<details>
<summary>Diff for <strong>main-HASH.js</strong></summary>

Diff too large to display
</details>
<details>
<summary>Diff for <strong>app-page-exp..ntime.dev.js</strong></summary>

Diff too large to display
</details>
<details>
<summary>Diff for <strong>app-page.runtime.dev.js</strong></summary>

Diff too large to display
</details>

</details>

</details>
Commit: 764b5c6f662da86c5401fb695f37112d4e6c784a